Update kubectl Installer for NS Mode & Doc Change
Added a NAMESPACE_MODE environment variable to the kubectl installer.
This env var configures the 'namespace_mode' variable in the Ansible
installer, therefore determining which Cluster Roles are installed to
enable the desired namespace functionality.

Also includes an update to the headers describing the various namespace
modes in the namespace documentation.