Updates installer from review and bug fix
This commit adds documentation on how to update the installer service account to
user image pull secrets. It also adds new inventory parameter to defaults in the
pgo-install.sh script. This new value was added to the inventory file and
ansible defaults but not here. This change syncs the two defaults. The parameter
lists are reordered to be alphabetical.