Remove boolean result from kubeapi.GetPVC
The value was ambiguous and in many cases it was used even if the API
call failed. A new function, GetPVCIfExists, returns a nil pointer when
the PVC is not found. Callers that only care if exists can check for nil.

Also renamed operator/pvc.Delete to DeleteIfExists because it checks for
NotFound first.