Allow WAL volume to be specified at cluster create
PostgreSQL uses write-ahead logging to ensure data integrity.[1] The
performance of the disk and file system of that log directly affect the
performance of PostgreSQL. Advanced deployments will want to separate
the WAL volume from the heap.

This implementation allows WAL storage to be specified only once during
cluster create. That spec carries through to replicas and clones. WAL
storage is entirely optional; when not specified WAL resides within the
PGDATA directory and volume.

There is an operator configuration to specify WAL storage on all newly
created clusters. When in use, the size of the volume can be overridden
per-cluster.

This implementation does not define the WAL volume in any deployment
templates because the volume name and mount path are constant.

[1]: https://www.postgresql.org/docs/current/wal.html

Issue: [ch6688]