Modify password generation functions to use crypto randomness
The previous random password generator was using the standard random
number generated (with a timestamp-based seed) and a character set
that included only mixed-case alphanumeric characters and numbers.

This replaces the password generation function to use the crypto/rand
library and extends the characters that can be used to include the
"usable" ASCII characters (code 33-126). Additionally, this replaces
some of the other "random generation" functions by using equivalent
ones found in the Kubernetes rand library.