Adds script to setup pgo client for local use
The ~/.pgo directory was setup by the ansible installer when it was run locally.
Now that the ansible is run in a container it won't be able to generate these
files so that the client can connect to the apiserver. This update adds a script
that pulls the relevant information from kubernetes secrets and generates the
files locally. It also provides feedback to the user if there is an error and
how to setup their environment so that the pgo client knows where the files are.