Support Graceful Shutdown of Worker Queues
With this commit, any/all worker queues are now gracefully shutdown.
Therefore, in the event that the Operator is shutdown (e.g. via a
SIGTERM sent when the pod running the Operator is terminated), this
provides an opportunity for any active work initiated via a work queue
to complete and/or cleanup prior to termination of the Operator.