Adds CCPImagePrefix and PGOImagePrefix to the pgcluster CRD
Previously, the CCPImagePrefix and PGOImagePrefix values
were set only in the global configuration. This update adds a
'CCPImagePrefix' and a 'PGOImagePrefix' value to the pgcluster CRD
which will be used in place of the global configuration values when
set. If not set, the existing behavior will be maintained and the
global configuration values will be used.