Add CCPImagePrefix to the pgcluster CRD
Previously, the CCPImagePrefix value was set only in the
global configuration. This update adds a 'CCPImagePrefix'
value to the pgcluster CRD which will be used in place of
the global configuration value when set. If not set, the
existing behavior will be maintained and the global configuration
value will be used.