Installer Metrics update: ns default and bugfix
This change adds in a default namespace for the metrics deployment. The default
will now be pgo if you use the installer image. The default namespace is also
updated in the inventory and ansible defaults for pgo-metrics.

This change also makes a bugfix where the fsGroup wasn't be set on the grafana
and prometheus deployments. This caused a permission error when the processes
tried to write to a required directory.