Removes ansible to setup installer image
This commit removes the ansible roles that were used to setup and run the
installer image job. This is now done through the kubectl client. The installer
image options are removed from the inventory file.

The docs are also updated to only have one page for the installer image. This
docs page specifies how to use `kubectl apply -f postgres-operator.yml`