Updates Client Setup Script: bugfix
This update changes how the setup script generates the `pgouser`, `client.crt`,
and `client.key` files. Currently new data will be appended to any existing
data. Now the script will overwrite the old data in each file.

The script now relies on the `go-template` output feature of `kubectl`. This
simplifies the script by decoding the username and password and generating the
`pgouser` file in one line. The commands to pull the client key and cert are
also updated to use `go-template`