Handle errors consistently in cluster and replica create
These two code paths are effectively very similar but differed in their
arrangement. It's now easier to compare them and implement changes that
affect both.

In both cases, errors from the unexported functions are returned and
checked before their caller publishes the result as an event.

A later commit will do more to align the creation of PVCs and volumes.

Issue: [ch6688]