Run pods with the process groups required by tablespaces
We allow every storage specification to include a list of group IDs that
a process needs to access that storage. When multiple types of storage
are mounted to a pod, processes within need the group IDs of _all_ those
specifications. Specifically, PostgreSQL needs access to the data
directory and all tablespaces.

This commit adds a function, pvc.CreateMissingPostgreSQLVolumes, to
handle this consistently across cluster create, scale, and restore. It
also removes kubeapi.CreatePVC and kubeapi.DeletePVCs which are no
longer used.

Issue: [ch6688]