Tie CPU Limits to CPU Requests
c9059ab8 and several commits after removed the ability to set container
resource limits on both memory and CPU for the PostgreSQL, pgBackRest,
and pgBouncer containers. While there are documented reasons as to why
we do not want to rely on Kubernetes treatment of the container memory
limit setting, the same concerns do not apply to CPU limits.

As such, this patch reinstates the ability to set CPU limits, but ties
it to the specific resource request. As such, making a request for CPU
resources simultaneously sets the request and limit parameters for the
container within the Pod. The default behavior does not change: no
CPU resources are requested for any Deployments.

Issue: [ch7919]