Add pgBouncer information to the cluster spec
Historically, pgBouncer enablement for a cluster resided in a label on
the pgcluster CRD, which gave it a relative impermanence relative to its
importance in a PostgreSQL cluster.

This patch moves pgBouncer to become first-class in the definition of a
PostgreSQL cluster, and introduces a "pgBouncer spec" that allows a user
to set if a pgBouncer is enabled and any resource definitions that it
should have. This makes a modification to how 43f5dc8 was implemented
(and well, the whole pgBouncer implementation for this release) as now
the resource utilization definition is stored within the PgStorageSpec
sub-struct.

This introduces a few noticeable changes and behaviors to how one can
manage a pgBouncer integration:

- Any changes to the pgBouncer portion of a custom resource
instantiation have full declarative behavior. In other words, one can
edit the custom resource directly and the changes will be applied

- The pgBouncer functionality is less reliant on the pgtask CRD. There
are a few more changes that need to occur in future commits, but the
direction is that pgBouncer updates will either occur imperatively
or declaratively without an intermediaries.

- While the update behavior works for modifying resource definitions,
there is a "double update" that presently occurs due to the transition
to utilizing a declarative API. This will be modified in a near-future
commit.

- The "uninstall" behavior that removes the "pgbouncer" role owned
objects from the target PostgreSQL database is disabled in this
commit, but will be re-enabled in a future commit and occur
impreatively.