Make the "uninstall" operation when deleting pgBouncer imperative
This moves the logic for "uninstall" out of the declarative portion of
the pgBouncer Operator code and into the API server, as this is an "ops"
decision instead of structural. By default, when the pgBouncer
Deployment is requested to be removed, the pgBouncer user is disabled
and the object artifacts remain.