Update StorageOS PVC Default Size
Currently, the default PVC size for StorageOS is set to
'300M'. This value is not interpreted properly and results in
a default size to be used. Additionally, 300MB is to small to
support even most basic testing deployments of PostgreSQL.

The PVC values throughout have been updated to '5Gi', which is both
recognized properly by StorageOS and is a more reasonable value
for testing that can easily be updated to a larger value as needed.