Allow for pgBouncer Deployments to be scaled
To deploy pgBouncer in a highly-available manner, one needs to run more
than one pgBouncer Pod across different Kubernetes nodes.

This extends the cluster CRD to container a value called "Replicas" in
the pgBouncer spec. Replicas de facto replaces "Enabled" as any
combination of < 1 replicas would effectively disable a pgBouncer
deployment (and vice versa). To follow good pattern, this attribute can
be modified on the cluster CRD and appropriately react to it, setting
the correct number of pgBouncer Pods in a Deployment.

This also introduces a few new commands to help with utilizing this new
feature:

- `pgo create cluster --pgbouncer-replicas`
- `pgo create pgbouncer --replicas`
- `pgo update pgbouncer --replicas`

Issue: [ch7921]