This commit enhances cluster configuration capabilities within the PostgreSQL Operator. Specifically, the '<clusterName>-pgha-config' configMap created for each cluster now manages the user-facing configuration settings available for a PG cluster.  This includes the cluster-wide configuration settings available via the Distributed Configuration Store (DCS), as well as the local configuration settings available for each individual database server (i.e. the primary and all replicas) within a PG cluster.
As clusters are created and replicas are added/removed, the
'<clusterName>-pgha-config' configMap is populated with the DCS
configuration and the local configuration for each server within the
cluster.  These configuration settings can then be modified by users as
needed (e.g. using 'kubectl edit cm <clusterName>-pgha-config'), after
which those settings will be applied to the cluster's DCS and/or any
local servers as appropriate.  And as servers (i.e. replicas) are
removed from the cluster, the configMap is properly cleaned up to
ensure any configurations that are no longer needed are removed.

In support of this change, a new ConfigMap controller has been created
that listens for configMap events within the Kubernetes cluster.
Specifically, the controller monitors for updates to
'<clusterName>-pgha-config' configMap's, and then executes sync logic
to ensure any configuration within that configMap is properly applied
to the associated PG cluster as needed.  Concurrency is utilized when
executing sync logic to concurrently sync both the cluster's DCS as
well as the local configuration for all servers within the cluster when
an event is received.

The ConfigMap controller is designed to be resilient to failures, e.g.
missed events and other transient failures.  Specifically, a refresh
interval is included that provides updates for all configMaps at a
regular interval (by default every 60 seconds), ensuring the controller
continuously tries to bring the PG cluster's configuration inline with
the configMap's configuration. Therefore, in the event that an event is
missed or an associated sync attempts fails, another event will be
picked up and processed during the next refresh interval, providing
another opportunity to properly sync all configuration (a re-queuing
strategy is also provided to capture transient errors encountered during
a sync attempt, providing a means to reattempt an event, leveraging
work queue rate limiting leveraged as needed).

Also, in support of quickly and effectively processing configMap
events, the configMap controller currently leverages two workers, which
are both able to concurrently listen for and process events placed on
the controller's worker queue. Additionally, the controller manager has
now been updated to ensure the caches for all informers created for use
by various controllers (including those outside of the ConfigMap
controller) are fully synced prior to running those informers.

To support the syncing of both DCS and local configuration within a PG
cluster, a new 'operator/config' package has been created, which
contains the resources and logic needed to manage both forms of PG
cluster configuration, i.e. both DCS and local server configuration.
Both forms of configuration share similar semantics for managing
configuration.  This includes the ability to "apply" specific
configuration as stored in the '<clusterName>-pgha-config' configMap to
a target resource (i.e. the DCS or a local server), the ability to
"refresh" the configuration stored in the '<clusterName>-pgha-config'
configMap using the configuration currently found in a target
resource (i.e. the DCS or a local server), and the ability to update
a configuration setting within the configMap (which includes deleting
unneeded configurations).  The local DB configuration logic is also
designed to update all servers within a cluster concurrently, avoiding
bottlenecks when updated the local configuration across a large PG
cluster.

A Sync() function is available for each form of configuration, which
can be called by the configMap controller when update events are
received to ensure the appropriate apply/refresh/update logic is
executed to bring the configuration within the configMap inline with
the cluster.  Update functions are also available to provide clients
(e.g. the API server) with an easy means of updating a specific
configuration (i.e. the DCS or a local server). Ultimately this concept
can be extended to include other forms of configuration for a PG
cluster that require sycning between a configMap and a target
application, service, system, etc.

Also includes updates to the scheduler as needed to align with
controller changes (e.g. the addition of logic to wait for
informer caches to sync), as well as updates to the dep lock file in
support of additional packages utilized for the configuration
capabilities described above.  Controller refresh intervals (which
right now are only utilized with the namespace and configMap
controllers) are exposed via settings in the 'pgo.yaml'.  While these
settings are not typically intended to be modified (with both
defaulting to 60 seconds), this allows for administrative fine tuning
of both refresh intervals if needed.