Scheduled Pgbackrest Container Prefix Passing
There is a bug where the correct image prefix is
not properly taken from the pgcluster CRD when creating
a pgbackrest schedule. The bug results in an empty value that
is not properly checked when submitting the associated pgtask.

This update corrects the initial error, and adds an additional
check that the value is set before submitting the task.