Update Makefile: use $HOME for $GOPATH
The makefile is currently setup to have a default GOPATH if the environment
variable is unset. This default is `/home/odev/go` instead of following the
documented example for GOPATH which is `$HOME/odev`. This change updates the
default to pull the `home` directory from the environment instead of relying on
the hardcoded value of `/home`. This will more closely follow the documented
example.