Cleanup & Add Configuration for Worker Counts
Includes general cleanup, as well as the addition of configurable
worker counts for those controllers that utilize worker queues.
Specifically, it is now possible to configure the number of workers
that will listen for items on the worker queues created for the
ConfigMap controller, PGCluster controller, PGReplica controller
and PGTask controller.