Installer: adds DEPLOY_ACTION to install yaml and directory update
This change adds the deploy action environment variable to the
`postgres-operator.yml` env block. This will follow the documentation more
accurately when it says to update the value. This also gives the user a better
idea of what the default action is without reading the docs.

As the installer has developed the `installers/method` directory is no longer
necessary. This change removes it.