Cleanup Namespace Validation
This commit cleans up namespace validation within the API server.
Within the 'ns' package, a ValidateNamespacesWatched func now
provides a single method of validating whether one or more
namespaces are watched by the current Operator installation.
Additionally, function 'GetNamespaceList' has been renamed to
'GetInitialNamespaceList' to indicate that it is responsible for
obtaining an initial list of watched namespaces during Operator
initialization, while function 'GetNamespaces' has been renamed
to GetCurrentNamespaceList to indicate that it obtains to current
list of watched namespaces post-Operator initialization.  Otherwise,
various other cleanup is included.

In support of these changes to the 'ns' package, the API server
functions for validating namespaces, i.e. that a namespace is
a watched namespace, have also been updated.  This ensures a
single, consolidated method for validating watched namespaces
is utilized by the API server, which is able to properly account for
the Operator's current namespace Operating mode when determining
whether or not a namespace is watched.

And finally, automated tests have also been updated as needed in
support of these changes, while the unneeded global env var for nslist
has been removed from the scheduler.