Orchestrate Repo & Primary DB Deployments
This commit updates cluster initialization logic to ensure the pgBackRest
repository is online and ready prior to initializing the primary database
container.  This ensures that the pgBackRest repository host is always
available and online as needed to fully initialize the database cluster,
e.g. to create the stanza, initial backup, replicas, etc.  This is
especially beneficial in environments where latency can lead to delays in
pulling and initializing the pgBackRest repository image, therefore
leading to database initialization (and therefore attempts to create the
stanza, etc.) prior to the repo coming online.

To ensure the repo is always initialized prior to initializing the database,
the primary deployment is now initially created with 0 replicas.  Then,
once the Operator detects that the repo host is ready (specifically via a
new repo init handler in the Pod controller), the primary database is
scaled to 1.  From here cluster creation then proceeds in the same manner
as before.

It should be noted that with this change the repo host also initially started
with 0 replicas, but then is immediately scaled up.  This is to ensure
that all Deployments and supporting resources (services, configMaps, PVCs,
etc.) have been created and are ready to go prior to creating any Pods. This
prevent potential race conditions, e.g. the repo host being created and
becoming ready prior the creation of the primary DB deployment.  With this
solution we can be sure the primary deployment always exists by the time
the repo host becomes ready and attempts to scale it to continue initialization
of the cluster.