Upgrade Documenation Updates and Cleanup
This update removes older versions of the manual
upgrade procedures, corrects documentation links where
needed and adds clarifying steps for manually upgrading
clusters to the latest version. This is needed for clusters
that have a primary PVC that is different from the default
cluster name, which  can occur after a failover or a restore.

Co-authored-by: jose-joye jose-joye@users.noreply.github.com