Allow for user to specify memory limits for certain containers
With the work on memory management for this release, one of the
discoveries made involved how the OOM killer on Kubelets would interact
with PostgreSQL instances (as well as pgBackRest, pgBouncer) in ways
that could lead to some undesired behavior. As such, the ability to
implicitly set memory limits was removed for the time
being.

However, there may be cases where one may want to set a memory limit,
along side a memory request + CPU request/limit to get a Guaranteed
QoS on a Pod, perhaps for pgBouncer or pgBackRest. This patch allows
for this via the `--enable-memory-limit` flag, specifically:

  pgo create cluster --enable-memory-limit
  pgo create cluster --enable-pgbackrest-memory-limit
  pgo create cluster --enable-pgbouncer-memory-limit
  pgo update cluster --enable-memory-limit
  pgo update cluster --disable-memory-limit
  pgo update cluster --enable-pgbackrest-memory-limit
  pgo update cluster --disable-pgbackrest-memory-limit
  pgo create pgbouncer --enable-memory-limit
  pgo update pgbouncer --enable-memory-limit
  pgo update pgbouncer --disable-memory-limit

The net effect is that if a memory request is present (and all of
PostgreSQL, pgBackRest, and pgBouncer have default values for memory
requests), the limit will also be set if it is enabled. By default, the
PostgreSQL Operator does not setting the memory limit per the behavioral
issues mentioned above.

If the memory limit is enabled, it is set to be identical to the memory
request.

This setting is also reactive, i.e. if you modify it in the CR directly,
the PostgreSQL Operator will roll out the changes to the managed Pods.

Issue: [ch8043]