Ensure one can add tablespaces on an upgraded cluster
The "tablespaceMounts" portion of the CR would be "null" on an
upgrade, which in turn would cause the unmarshaled map to be
"nil". By initializing the map when a tablespace is being added
to a cluster that was ugpraded from an earlier version of the
PostgreSQL Operator, one can ensure successful addition and
execution.