moves kubeapi library folder into internal
This library is being moved into internal as step toward
organizing the postgres-operator libraries. This reduces
clutter in the top level project directory and supports
future reorganization, refactoring, and movement to Go
modules where appropriate.

At this time, all libraries are believed to have postgres-operator
specific code and are not likely to be dependencies of code outside
of postgres-operator.

However, should this be in err, each library move is split into a
separate commit for easy reversion.