Updates Installer Options and bugfix
Values that were included in the postgres-operator.yml file but not operator
configuration options (HOME and ANSIBLE_CFG) are now set in the deploy script.
These settings should not need to be updated by a user.

This change updates values in the pgo-deployer files that were incorrect. This
includes docs, the postgres-operator.yml deployment file, and the
inventory_template.