Move Postgres MD5 generation functions to internal
While the MD5 password generation was consoliated, it effectively limited
us to only supporting MD5. With the new interface, the PostgreSQL Operator
is set up to support other methods (e.g. SCRAM).

This also puts the code in the "internal" folder and adds some testing,
which after a certain commit point will not be news.