Allow for SCRAM hashed passwords to be set on pgo create/update user
This commit introduces the `--password-type` flag on the
`pgo create user` and `pgo update user` command that allows a user to
specify the following passwords types:

- "", "md5" => md5
- "scram", "scram-sha-256" => SCRAM-SHA-256

In turn, the PostgreSQL Operator will hash the passwords based on the
chosen method and store the computed hash in PostgreSQL.

When using SCRAM support, it is important to note the following
observations and limitations:

- When using one of the password modifications commands on
`pgo update user` (e.g. `--password`, `--rotate-password`, `--expires`)
with the desire to keep the persisted password using SCRAM, it is
necessary to specify the "--password-type=scram-sha-256" directive.

- SCRAM does not work with the current pgBouncer integration with the
PostgreSQL Operator. pgBouncer presently supports only one
password-based authentication type at a time. Additionally, to enable
support for SCRAM, pgBouncer would require a list of plaintext passwords
to be stored in a file that is accessible to it. Future work can
evaluate how to leverage SCRAM support with pgBouncer.

Issue: [ch8179]