Ensure client-setup.sh does not error on certain existing installs
An earlier version of the PostgreSQL Operator installer would set
file ownership to 0400 on certain files, which would cause issues with
the current client-setup.sh when it would attempt to overwrite these
files.

As the files are owned by the user executing the client setup script,
a solution is to relax the permissions on the files in question to
0600 (which is what they should be in their final form) before
attempting to overwrite them.

Issue: [ch8076]