Introduce the Restart Endpoint & Reload Cleanup
This commit introduces the 'restart' endpoint into the Operator REST
API, along with an associated 'pgo restart cluster' command in the
pgo client.  This allows users to perform a PostgreSQL restart on
one or more instances (i.e. primary or replicas) within a PG cluster.
Specifically, users can either restart all instances at the same
time using the following command:

pgo restart mycluster

Or one or more targets can be specified using the --target flag
(following similar semantics to the 'failover' and 'scaledown'
commands):

pgo restart mycluster --target=mycluster-abcd

The restart itself is performed by calling the Patroni 'restart'
REST endpoint on the specific instance (primary or replica) being
restarted.

As with 'failover' and 'scaledown' commands it is also possible
to specify the '--query' flag to query instances available for
restart.

pgo restart mycluster --query

And the output will display whether or not a restart is needed via
a "PENDING RESTART" column:

Cluster: mycluster2
INSTANCE                ROLE     ...  PENDING RESTART
mycluster2              primary  ...             true
mycluster2-ambq         replica  ...             true

It should be noted that the '--query' flag for the 'failover' and
'scaledown' commands now also includes the "PENDING RESTART"
information, which is now returned with any replication information
(the primary can now also be obtained when retrieving replication
information in support of the 'restart' use case).

Also included in this commit are automated end-to-end tests that test
the following restart capabilities:

- Querying for restart targets
- Restart all instances in a cluster
- Restarting a specific instance within a cluster using the --target
  flag

And finally, this commit also updates the 'reload' endpoint.
Specifically, running the 'pgo reload' command now properly reloads
all instances (i.e. the primary and all replicas) within the cluster.
Also, if more than one cluster is specified for a reload attempt, and
an error is encountered while attempting a reload for any of those
cluster, the endpoint will simply capture the error and move onto the
restart attempt for the next cluster, instead of abruptly halting all
processing and immediately returning an error.  The error message
returned then includes one or more errors depending on the number
clusters that encountered errors during the reload attempt.  While
ideally multiple errors would be returned in this scenario (instead of
multiple errors combined in a single error message, each separated by a
new line as seen here), this allows the 'reload' endpoint to react
properly when errors are encountered (i.e. not abruptly halt processing
on the first error)  without breaking the current API contract for the
'reload' endpoint (which defines a single error returned in an error
scenario).