Fix S3-Only Cluster Cloning
All S3 environment variables are now properly set on both the
pgBackRest repo Pod and the pgBackRest restore Job when cloning
a pgcluster.  This ensure the restore is able to complete
successfully during the second step of the clone workflow when
cloning a cluster that utilizes S3 only with pgBackRest.