Allow for FSGroup hardcode to be disabled
The FSGroup hardcode is meant to help people running the PostgreSQL
Operator in Kubernetes to be able to utilize the projected volume
mount of the TLS Secrets when using the TLS features of PostgreSQL.
This is needed to ensure that the keyfiles are both readable by
PostgreSQL (i.e. no more permissive than a 0640 permission), as
PostgreSQL will refuse to start if the TLS files are too open.[1]

However, in environmetns that support a restricted Security
Context Constraint (e.g. OpenShift), and, in particular, without
any SCC modifications, having the hardcoded FSGroup could cause a
PostgreSQL pod to refuse to start.

This patch introduces the ability to disable the FSGroup from
being referenced with the "DISABLE_FSGROUP" parameter. Setting
this to true will remove the FSGroup setting and allow for the
environment to dictate how to handle this. By default, it is set
to false, as Kubernetes needs to have the FSGroup present for
PostgreSQL to properly boot up with TLS keys present.

[1] https://www.postgresql.org/docs/current/ssl-tcp.html

Issue: [ch8235]