Decompose Deployer RBAC & Role Binding Cleanup
This commit creates a new ClusterRole specifically for use by the
'pgo-deployer-sa' ServiceAccount.  With this new role all privileges
required for the deployer container to successfully perform a
PostgreSQL Operator installation are now explicitly defined in the
'pgo-deployer-cr' ClusterRole (which is located in the
'postgres-operator.yaml' specification). This means the deployer
container, and therefore installing via 'kubectl', no longer requires
the use of "cluster-admin".

Also, various ClusterRoleBindings/RoleBindings that were recently
updated to use "kind: ServiceAccount" have been updated to remove
the "apiGroup", ensuring the bindings can be successfully created.