Set default limits on the collect sidecar
While it is possible to deploy a PostgreSQL cluster with a guaranteed
Pod QoS, adding in a metrics sidecar would cause the QoS to be changed
to burstable.

While the future may hold for allowing the user to tune the metrics
sidecar from the CRD/API etc., the short term solution is to set
sane defaults for deploying the metrics sidecar and to allow for
people to modify the existing template to the values that they wish
to have in it.

Issue: [ch8253]
Issue: #1526