Allow for FSGroup for pgAdmin to be disabled via DISABLE_FSGROUP
adea5676d introduced the ability to disable the FSGroup from being set
on the PostgreSQL related Pods, which in turn would enable deployments
under the OpenShift Restricted SCC. However, the pgAdmin Pod sets its
own FSGroup of 2, and in order to deploy it under this SCC, also
requires the option to disable the FSGroup from being explicitly set.

This patch extends the functionality of adea5676d to the pgAdmin Pod.

Issue: [ch8254]